refactor(cli): align group command modules with session
