+++ apps/cli/src/commands/group/options.ts
@@ -1,0 +1,7 @@
+import type { StationConfig } from "@station/config";
+
+export type GroupCommandOptions = {
+  config?: StationConfig;
+  configPath?: string;
+  timeoutMs?: number;
+};
